Prevent System.exit(int) and sys.exit() from being called (#12)
* Prevent System.exit() from being called

This is achieved by excluding the method from being added to the dictionary.

* Prevent sys.exit() from being called in embedded mode

* Move test to org.python.core

* Add a test verifying the blocking of java.lang.System.exit()

The test asserts that System.exit(int) is prevented in embedded mode.
However, we generally block it, affecting also console mode.

There you can still use:
>>> import sys
>>> sys.exit()

* Fix test_cmd_line, which calls sys.exit()

* Be a bit less restrictive, to let test_cmd_line_script still pass

It should be possible to call runpy and report errors as messages back.